requirements.txt
anyio==4.12.1
APScheduler==3.11.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cryptography==46.0.5
google-auth==2.49.0
google-auth-oauthlib==1.3.0
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
oauthlib==3.3.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.2.2
python-telegram-bot==22.6
pytz==2026.1.post1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
tzlocal==5.3.1
urllib3==2.6.3
